Connection: revert most of previous commit
Ensuring proper 2-way TLS shutdown on service shutdown is not feasible,
it requires to delay shutdown and won't work if close is never called on
the connection.

Instead, just ensure 1-way TLS shutdown when a connection is immediately
destroyed. A consumer can always have the full 2-way shutdown by
explicitly closing a connection.