Config: refactor to make args parser stackable
So far, the args parser *always* created a new PSC_Config object for any
subsection. This doesn't hurt when it's the only parser, but is *wrong*
otherwise, it should only do that for a list of subsections, so
different parsers can stack nicely (overriding single values even in
subsections and only appending to lists). Fixed that...