Service: Add function validating fd
Add a function to validate whether a given fd is usable with select().
Use this everywhere fds are created to register them for reading/writing
as a safeguard to prevent unexpected/undefined behavior when too many
fds are watched.

Also add build-time configuration to set a desired FD_SETSIZE for
systems supporting this, set the default to 4096.