ThreadJob: Make stack size configurable
Add a method to configure the size of a job's private stack. Also defer
allocating that stack until the job is actually executed.

Add a static PSC_AsyncTask method to check whether awaiting a task will
always block a worker thread (because support for POSIX user context
switching is missing).