Connection: Try writing without notification
Don't directly register for "ready to write" notifications when
attempting to write something. Instead, when there's something to write
at the end of the event loop, try just doing that and only register for
notification when not everything could be written.

This should reduce syscall overhead because the more common situation is
that there's room in the system's send buffer.