Server: Use internal event for shutdown
Instead of keeping a dynamic array with active connections, only keep
the total number and add a private Event to close them all on shutdown.

Also remove the clientDisconnected() event, it only consumes RAM and
CPU while consumers could also subscribe to the closed() event of the
individual connections to get the same information.