ThreadPool: Refactor for better performance
Remove complex logic to put work on pool threads, replace with a single
MPMC queue where every thread job is enqueued and all pool threads wait
on a shared semaphore for available jobs. This requires to allow
resuming a job after awaiting an AsyncTask on a different pool thread.

Also replace the legacy "timeoutTicks" with a value in milliseconds.