Connection: Add pool to allow reusing memory
Add an internal ConnectionPool class, maintaining a list of active and
available connection objects, and allow passing a pool to the internal
connection constructor, so a new connection will only be allocated if
there isn't an unused one left in the pool.

Use this in Server code to have one pool per thread for maintaining
client connections. This also allows a simple way to close all remaining
connections on forced shutdown (destroying the server).

Therefore, also remove the hack to allow closing a connection from a
different thread, this is not needed any more.