Log: Fix async logging with service workers
Fix async logging with multiple service workers. Enqueueing a thread job
from within a pool thread is not safe any more, because the job must
know its "owning" service worker thread. So, check whether we're logging
from within a pool thread, and if so, delegate creating and enqueuing
the thread job to the main thread.