Server: Add simple load balancing
When distributing accepted connections to service worker threads,
replace the dumb round-robin by a simple load-balancing algorithm:

For each thread, maintain a counter of currently opened connections.
When looking for a thread to assign a new connection, search for the one
with the lowest number of connections, starting at the thread that was
used last.